initial version

+++ src/main/groovy/groovy/blog/TodoStats.groovy
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.annotation.JsonCreator
 import com.fasterxml.jackson.annotation.JsonProperty
 
 @JsonCreator
-record TodoStats(@JsonProperty('total') int total,
-                 @JsonProperty('completed') int completed,
-                 @JsonProperty('totalScheduled') int totalScheduled,
-                 @JsonProperty('completedOnSchedule') int completedOnSchedule) {}
+record TodoStats(@JsonProperty('total') long total,
+                 @JsonProperty('completed') long completed,
+                 @JsonProperty('totalScheduled') long totalScheduled,
+                 @JsonProperty('completedOnSchedule') long completedOnSchedule) {}
